requirements.txt
# Python Dependencies for Retail Layout Optimizer
# Pin versions for reproducibility

# Data processing
pandas==2.1.4
numpy==1.26.2

# Machine learning
scikit-learn==1.3.2
scipy==1.11.4

# Graph analysis
networkx==3.2.1
python-louvain==0.16

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0

# Web framework
Flask==3.0.0
Flask-Cors==4.0.0

# Visualization
plotly==5.18.0

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code quality (development only)
black==23.12.1
ruff==0.1.8